collector: use urllib to replace urllib2
urllib2 will raise an exception when the resouce is not found,
the exception can not be serialized by python, urllib do not raise
exception and just return the content of a body which is expected.

The result shoulb be serialized because we use the process poll go
run every collector.