api: server web and deal with frontend route
The frontend use fake router to cheat browser, however, if user
refresh the page or copy the url to another browser, lies will
exposed with an unfriendly 404 page. To help frontend keep the
secret, the api will return the index.html if it meet 404 with
path not start with /api/v1.