feature: introduce remote log gc (#79)
If the cluster has many nodes, or the cluster run a long time
without cleanup logs, the logs will stack and fill up the disk
of the machine where foresight running.

This commit introduce the gc mechanism, where the log size will
be checked in a specific interval (configurable). If the log size
exceed the limit, the gc will truncate the oldest ones to make
them empty. Since we use --update flag on rsync, the rsync will
not cover them again because they are newer than remote.

Signed-off-by: lucklove <gnu.crazier@gmail.com>